Fix warning highlighting.
So far, we just did a normal, incremental cabal build,
which resulted in warnings only being generated once
and vanishing with the next save.

This adds a configurable new option which is now the default:
After running an incremental build, we can recompile everything
with -Wall (and -fno-code to make it faster).

This results in errors being shown fast as usual and code being built,
and warnings popping up slightly afterwards.

This is the new default for both autosave and the build command.

There are also some new options to run a "typecheck" build
that is faster by leaving out as much codegen as possible
(currently this is just -c for not linking).