better environment var replacements
Bug: add_to_PATH didn't actually do anything because subprocess.Popen()
did not use the augmented PATH to find the executable. It only passed
the augmented environment to the subprocesses.

There are two options:

(a) Temporarily override os.environ['PATH'] so that subprocess.Popen()
    can actually find the executable.

(b) Actually find the executable.

Option (b) is currently used.

FIXME: Need to check the return from call_and_* so that the caller can
DTRT if the program isn't found or installed.