deprecate ST2
Completely deprecate SublimeText 2.  This plugin is hideously broken for
ST2, so better to say a fond farewell than fighting a losing fight.
Even when imports are cleaned up, there are still a lot of calls into
the ST3 API that just can't be fixed.

"The Highlander" (*) variables: status_message_manager, agent, client,
client_back: Ensure that they aren't overwritten or reinitialized when
the plugin reloads.

(*) "There can be only one."