import module fixed, autocompletion debugging
Fixed "Add module"/"Insert import for symbol", adding the import
after the module's "where" if there are no imports. The import is
only added to the end of the file if all other alternatives fail.

Tried to address autocompletion issue: new function names do not show
up in the completion list (until after the file is re-scanned.)