why multiple hsdevs?
Trying to track down why multiple hsdevs start up from time to time,
as well as generally improve the backend lazy loading. Currently,
there's a lock in front of the major backend initialize() and
shutdown() methods, which should alleviate the problem.

  -- Consolidated some of the logic between backend manager state
     transitions, which should make the code a little less error prone.

  -- Backend manager has context manager __enter__ and __exit__,
     which should also make the code more robust.

Found a minor issue with "SublimeHaskell: Run": which() didn't check
if the command is an absolute path. If given an absolute path, which()
returns successfully.