more cleanups
Looked into the SourceHaskellTypeCache and it really doesn't do much,
unless the backend is hsdev. Might move that into hsdev. Doesn't do
anything for ghc-mod.

Also, need to reduce the need to call hsinspect, unless the backend
is (again) hsdev. But, to do that, need to be able to read cabal
files.