clean imports refactoring
"Clean Imports" (aka "Clear Imports") command refactored so that
'hsclearimports' is invoked by the 'hsdev' backend. This is also
takes care of cabal/stack exec wrapping.

ProcHelper.invoke_tool is now gone. No more call sites, due to
all of the refactoring into the backends. Don't need to know if a
tool is available since it's now specific to the backend.