backend options set in a way that should succeed
The 'cabal-new-build' backend example was configured in a way that
should be accepted, but wasn't:

        "Cabal new-build local hsdev": {
            "default": false,
            "backend": "hsdev",
            "options": {
              "local": true,
              "port": 4567,
              "exec-with": "cabal-new-build"
              // ^^^^^^^^^^^^^^^^^ Remember to insert a comma (",") after this line once you set install-dir
              // to hsdev's installation directory.
              //
              // Force a diagnostic message because this has to be set in conjunction with exec-with:
              // "install-dir": "$HOME/totally/invalid/directory"
            }

Note the '"local": true' line: technically, true, but not accepted and
eventually causes a Python traceback because install-dir isn't set.

Update the tests to make sure that exec-with and install-dir are both
set, even if local is true. Mark the backend unavailable if either
exec-with or install-dir are missing.