issue 377, 373, 359: incomplete reads from hsdev client socket
Solve issue 377 (and possibly issues 373 and 359) -- @coodoo noticed
a UnicodeDecodeError traceback after saves, which turns out to be a
short read from the socket -- more data is available, which
SublimeHaskell should continue to read in order to ensure that
decode('utf-8') succeeds. The fix here is to continue reading from
the socket if decode('utf-8') raises the UnicodeDecodeError exception.
If no exception is raised, then split and process what looks like a
complete request.

This probably also fixes issues 373 and 359, which manifested similar
behavior, but without the traceback.