make fly mode great again
Reimplement fly mode, using a ViewEventListener class that's specific
to Haskell source. This avoids some of the previous gymnastics related
to which view was currently active and being checked in fly mode.

Refactor the event listener into view-specific event listeners. There is
a workaround for some ST3 Python API wackiness.

  -- One would think that inheritance might work to reduce complexity,
     but the ST implementation of ViewEventListener doesn't call all of
     the Python machinery to make inheritance work cleanly.

  -- Common code lives in its own class and there are "redundant"
     methods (like on_new) in the child classes that call the common
     base class' code. Each child class has its own "on_new" that calls
     the base class implementation. The base class should implement its
     own version of "on_new", but can't because ViewEventListener never
     calls super().__init__(), so inheritance doesn't get fully
     initialized.