ghc-mod backend tweaks
It turns out that the ghc-mod backend's "error" marker is just
informative -- it doesn't mean that a command failed and has nothing
to do with the success or failure of a command. Consequently, just
log whatever shows up with the error prefix to an output panel, just
like Emacs does in the *GHC Error* buffer.

Tripped across this issue on a newer version of ghc-mod, which decided
to invoke stack and proceeded to build the project.