Prepare functions handling responses for unexpected types (#427)
The backend `call` function can return boolean True or False, where a list is
expected in normal circumstances. This can happen, for example, if the hsdev
process is killed. In such cases, the involved functions raise a TypeError.

This behaviour can be seen in the on-hover and completions handlers, upon
killing the hsdev process.